悬赏0.8

+++ foreground/src/main/java/com/pyteam/foreground/controller/Ad01Controller.java
@@ -110,7 +110,6 @@ public class Ad01Controller
         if(isLogin)
         {
             m.addAttribute("isLogin", isLogin);
-            System.out.println("？？？"+aac202);
             m.addAttribute("ad01", ae09Service.sel(aac202));
             return "select";
         }
@@ -118,6 +117,13 @@ public class Ad01Controller
         m.addAttribute("isLogin",isLogin);
         return "select";
     }
+
+    @GetMapping("/chon")
+    public void shuijiao(Model m,@RequestParam("aad101")int aad101)
+    {
+        m.addAttribute("ad01",ad01Service.findbyaad101(aad101));
+    }
+
 
     @PostMapping(value="/select")
     public String likeQuery(@RequestParam(value = "aad105" ,required=false) String question, HttpSession session, Model m)
@@ -133,16 +139,30 @@ public class Ad01Controller
     }
 
     @PostMapping ("/add")
-    public String add(@ModelAttribute (value = "ad01")Ad01 ad01,HttpServletRequest request, HttpServletResponse response)
+    public String add(@ModelAttribute (value = "ad01")Ad01 ad01,HttpServletRequest request, HttpServletResponse response,Model m)
     {
-        ad01.setAab101(Integer.parseInt(getCookies(request, "userId")));
-        ad01Service.add(ad01);
+        isLogin = isLogin(request, response);
+        if(isLogin)
+        {
+            m.addAttribute("isLogin",isLogin);
+            ad01.setAab101(Integer.parseInt(getCookies(request, "userId")));
+            ad01Service.add(ad01);
+            return "add";
+        }
+        m.addAttribute("isLogin",isLogin);
         return "add";
     }
     @GetMapping("add")
-    public String showadd()
+    public String showadd(Model m,HttpServletRequest request,HttpServletResponse response)
     {
-        return "add";
+        isLogin = isLogin(request, response);
+        if(isLogin)
+        {
+            m.addAttribute("isLogin",isLogin);
+            return "add";
+         }
+        m.addAttribute("isLogin",isLogin);
+        return "ad01list";
     }
 
     @PostMapping("/del")
@@ -160,11 +180,18 @@ public class Ad01Controller
         return "del";
     }
     @RequestMapping("/delad01")
-    public String delete(@RequestParam(value = "aad101",required = false) int id)
+    public String delete(@RequestParam(value = "aad101",required = false) int id,HttpServletResponse response ,HttpServletRequest request,Model m)
     {
-        System.out.println(id);
-        ad01Service.deleteById(id);
-        return "edit";
+        isLogin = isLogin(request, response);
+        if(isLogin)
+        {
+            m.addAttribute("isLogin", isLogin);
+            System.out.println(id);
+            ad01Service.deleteById(id);
+            return "edit";
+        }
+        m.addAttribute("isLogin",isLogin);
+        return "ad01list";
     }
 
     @RequestMapping("cxk")
